third_party/blink/web_tests/wpt_internal/css/css-ui/webkit-user-modify-shadow-dom.html

<!doctype html>
<title>-webkit-user-modify is reset for shadow root children</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #outer > div { -webkit-user-modify: read-write; }
</style>
<div id="outer">
  <div id="host">
    <template shadowrootmode="open"><i></i></template>
  </div>
  <div>
    <i id="light"></i>
  </div>
</div>
<script>
  test(() => {
    assert_equals(getComputedStyle(light).webkitUserModify, "read-write",
                  "Inherited from parent");
    assert_equals(getComputedStyle(host.shadowRoot.firstChild).webkitUserModify, "read-only",
                  "Reset after inherited from host parent");
  }, "Inherited -webkit-user-modify reset on shadow root children");
</script>
